# openapi.yaml
openapi: 3.1.0
info:
  title: Documan API
  version: 1.0.0
  description: |
    REST API for organising academic subjects and files, and for the surrounding community
    features (posts, comments, votes, favourites, users, roles).

    This document is maintained by hand rather than generated at runtime. When an endpoint,
    payload or status code changes, update it in the same commit.

    Notes on conventions:
      * Collection endpoints are paginated and return a `PageResponse` envelope.
      * Errors use RFC 9457 `application/problem+json`.
      * Identifiers are passed as query parameters.
      * There is currently no authentication; every operation is reachable anonymously.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /api/v1/user:
    get:
      tags: [User]
      summary: Fetch a user by id
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "200": { $ref: "#/components/responses/User" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [User]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
    put:
      tags: [User]
      summary: Update a user
      description: A null `password` leaves the stored credential unchanged.
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [User]
      summary: Delete a user
      description: Fails with 409 if the user still owns posts or comments.
      parameters: [{ $ref: "#/components/parameters/UserId" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/user/username:
    get:
      tags: [User]
      summary: Fetch a user by username
      parameters:
        - name: username
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/user/all:
    get:
      tags: [User]
      summary: List users
      parameters: [{ $ref: "#/components/parameters/Pageable" }]
      responses:
        "200": { $ref: "#/components/responses/UserPage" }
  /api/v1/user/posts:
    get:
      tags: [User]
      summary: Posts authored by a user
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/PostPage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/user/comments:
    get:
      tags: [User]
      summary: Comments authored by a user
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/CommentPage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/user/subjects:
    get:
      tags: [User]
      summary: Subjects for a user's department, year and semester
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/SubjectPage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/user/favourites/posts:
    get:
      tags: [User]
      summary: Posts a user has favourited
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/PostPage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/user/favourites/files:
    get:
      tags: [User]
      summary: Files a user has favourited
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/FilePage" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [User]
      summary: Favourite a file
      description: Idempotent; favouriting twice leaves the tally unchanged.
      parameters:
        - { $ref: "#/components/parameters/FileId" }
        - { $ref: "#/components/parameters/UserId" }
      responses:
        "200": { $ref: "#/components/responses/File" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/user/favourites/files/remove:
    post:
      tags: [User]
      summary: Remove a file favourite
      parameters:
        - { $ref: "#/components/parameters/FileId" }
        - { $ref: "#/components/parameters/UserId" }
      responses:
        "200": { $ref: "#/components/responses/File" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/user/votes/posts:
    get:
      tags: [User]
      summary: Posts a user has voted on, filtered by direction
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/VoteType" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/PostPage" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/user/votes/comments:
    get:
      tags: [User]
      summary: Comments a user has voted on, filtered by direction
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/VoteType" }
        - { $ref: "#/components/parameters/Pageable" }
      responses:
        "200": { $ref: "#/components/responses/CommentPage" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  parameters:
...
    UserId:
      name: userId
      in: query
      required: true
      schema: { type: integer, format: int32 }
...
    FileId:
      name: fileId
      in: query
      required: true
      schema: { type: integer, format: int32 }
...
    VoteType:
      name: voteType
      in: query
      required: true
      schema: { $ref: "#/components/schemas/VoteType" }
    Pageable:
      name: pageable
      in: query
      required: false
      description: |
        Standard Spring Data pagination. Supply as separate query parameters:
        `page` (0-based), `size` (max 100), and `sort` as `property,asc|desc` (repeatable).
      style: form
      explode: true
      schema:
        type: object
        properties:
          page: { type: integer, minimum: 0, default: 0 }
          size: { type: integer, minimum: 1, maximum: 100, default: 20 }
          sort:
            type: array
            items: { type: string }
  responses:
    User:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UserResponse" }
    UserPage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/UserResponse" }
...
    PostPage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/PostResponse" }
...
    CommentPage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/CommentResponse" }
...
    SubjectPage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/SubjectResponse" }
    File:
      description: OK
      content:
        application/json:
          schema: { $ref: "#/components/schemas/FileResponse" }
    FilePage:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/PageResponse" }
              - type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/FileResponse" }
...
    NotFound:
      description: The addressed resource does not exist
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    Conflict:
      description: Uniqueness, referential or optimistic-locking conflict
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetail" }
    ValidationFailed:
      description: One or more fields failed validation
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ValidationProblemDetail" }
...
  schemas:
    VoteType:
      type: string
      enum: [UPVOTE, DOWNVOTE]
    PageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, first, last]
      properties:
        content: { type: array, items: {} }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
    CreateUserRequest:
      type: object
      required: [username, password, firstName, lastName, email, departmentId, yearId, semesterId]
      properties:
        username: { type: string, maxLength: 40 }
        password: { type: string, minLength: 8, maxLength: 200 }
        firstName: { type: string, maxLength: 60 }
        lastName: { type: string, maxLength: 60 }
        email: { type: string, format: email, maxLength: 180 }
        departmentId: { type: integer, format: int32 }
        yearId: { type: integer, format: int32 }
        semesterId: { type: integer, format: int32 }
        acceptedTermsOfService: { type: boolean, default: false }
    UpdateUserRequest:
      type: object
      required: [username, firstName, lastName, email, departmentId, yearId, semesterId]
      properties:
        username: { type: string, maxLength: 40 }
        password:
          type: [string, "null"]
          minLength: 8
          maxLength: 200
          description: Omit or send null to leave the stored password unchanged.
        firstName: { type: string, maxLength: 60 }
        lastName: { type: string, maxLength: 60 }
        email: { type: string, format: email, maxLength: 180 }
        departmentId: { type: integer, format: int32 }
        yearId: { type: integer, format: int32 }
        semesterId: { type: integer, format: int32 }
        acceptedTermsOfService: { type: [boolean, "null"] }
        canPost: { type: [boolean, "null"] }
        canComment: { type: [boolean, "null"] }
...
    DepartmentResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
    YearResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        value: { type: string }
    SemesterResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
    RoleResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
    UserResponse:
      type: object
      description: The password is never included in a response.
      properties:
        id: { type: integer, format: int32 }
        username: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        acceptedTermsOfService: { type: boolean }
        verified: { type: boolean }
        canPost: { type: boolean }
        canComment: { type: boolean }
        role: { $ref: "#/components/schemas/RoleResponse" }
        department: { $ref: "#/components/schemas/DepartmentResponse" }
        year: { $ref: "#/components/schemas/YearResponse" }
        semester: { $ref: "#/components/schemas/SemesterResponse" }
        dateCreated: { type: string, format: date-time }
        dateLastInteracted: { type: string, format: date-time }
    PostResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        title: { type: string }
        description: { type: string }
        content: { type: string }
        upvoteCount: { type: integer, format: int64 }
        downvoteCount: { type: integer, format: int64 }
        favouriteCount: { type: integer, format: int64 }
        authorId: { type: integer, format: int32 }
        authorUsername: { type: string }
        dateCreated: { type: string, format: date-time }
        dateModified: { type: string, format: date-time }
    CommentResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        content: { type: string }
        upvoteCount: { type: integer, format: int64 }
        downvoteCount: { type: integer, format: int64 }
        postId: { type: integer, format: int32 }
        authorId: { type: integer, format: int32 }
        authorUsername: { type: string }
        dateCreated: { type: string, format: date-time }
        dateModified: { type: string, format: date-time }
    SubjectResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string }
        code: { type: string }
        lab: { type: boolean }
        theory: { type: boolean }
        department: { $ref: "#/components/schemas/DepartmentResponse" }
        year: { $ref: "#/components/schemas/YearResponse" }
        semester: { $ref: "#/components/schemas/SemesterResponse" }
    FileResponse:
      type: object
      properties:
        id: { type: integer, format: int32 }
        name: { type: string, description: Original client-supplied filename. }
        objectName: { type: string, description: Generated object key in the bucket. }
        objectURL: { type: string, format: uri }
        size: { type: integer, format: int64 }
        favouriteCount: { type: integer, format: int64 }
        subjectId: { type: integer, format: int32 }
        dateCreated: { type: string, format: date-time }
        dateModified: { type: string, format: date-time }
...
    ProblemDetail:
      type: object
      description: RFC 9457 problem detail.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        timestamp: { type: string, format: date-time }
    ValidationProblemDetail:
      allOf:
        - { $ref: "#/components/schemas/ProblemDetail" }
        - type: object
          properties:
            errors:
              type: object
              description: Field name to failure message.
              additionalProperties: { type: string }